remove QueryHelper
