+++ src/Query/Query.php
@@ -14,7 +14,6 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Helper\ArrayHelper;
-use Yiisoft\Db\Query\Helper\QueryHelper;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
 use function array_merge;
@@ -76,7 +75,6 @@ class Query implements QueryInterface
     private Closure|string|null $indexBy = null;
     private ExpressionInterface|int|null $limit = null;
     private ExpressionInterface|int|null $offset = null;
-    private QueryHelper|null $queryHelper = null;
     private array|string|ExpressionInterface|null $where = null;
 
     public function __construct(private ConnectionInterface $db)
@@ -112,7 +110,7 @@ class Query implements QueryInterface
 
     public function addOrderBy(array|string|ExpressionInterface $columns): static
     {
-        $columns = $this->createQueryHelper()->normalizeOrderBy($columns);
+        $columns = $this->normalizeOrderBy($columns);
 
         if ($this->orderBy === []) {
             $this->orderBy = $columns;
@@ -187,7 +185,7 @@ class Query implements QueryInterface
             return $this->select($columns);
         }
 
-        $this->select = array_merge($this->select, $this->createQueryHelper()->normalizeSelect($columns));
+        $this->select = array_merge($this->select, $this->normalizeSelect($columns));
 
         return $this;
     }
@@ -544,7 +542,7 @@ class Query implements QueryInterface
 
     public function orderBy(array|string|ExpressionInterface $columns): static
     {
-        $this->orderBy = $this->createQueryHelper()->normalizeOrderBy($columns);
+        $this->orderBy = $this->normalizeOrderBy($columns);
 
         return $this;
     }
@@ -645,7 +643,7 @@ class Query implements QueryInterface
 
     public function select(array|string|ExpressionInterface $columns, string $option = null): static
     {
-        $this->select = $this->createQueryHelper()->normalizeSelect($columns);
+        $this->select = $this->normalizeSelect($columns);
         $this->selectOption = $option;
 
         return $this;
@@ -771,17 +769,169 @@ class Query implements QueryInterface
         return $command->queryScalar();
     }
 
-    private function createQueryHelper(): QueryHelper
+    /**
+     * Removes {@see isEmpty()|empty operands} from the given query condition.
+     *
+     * @param array|string $condition the original condition
+     *
+     * @return array|string the condition with {@see isEmpty()|empty operands} removed.
+     */
+    private function filterCondition(array|string $condition): array|string
     {
-        if ($this->queryHelper === null) {
-            $this->queryHelper = new QueryHelper();
+        if (!is_array($condition)) {
+            return $condition;
         }
 
-        return $this->queryHelper;
+        if (!isset($condition[0])) {
+            /** hash format: 'column1' => 'value1', 'column2' => 'value2', ... */
+            /** @var mixed $value */
+            foreach ($condition as $name => $value) {
+                if ($this->isEmpty($value)) {
+                    unset($condition[$name]);
+                }
+            }
+
+            return $condition;
+        }
+
+        /** operator format: operator, operand 1, operand 2, ... */
+        /** @var string */
+        $operator = array_shift($condition);
+
+        switch (strtoupper($operator)) {
+            case 'NOT':
+            case 'AND':
+            case 'OR':
+                /** @psalm-var array<array-key, array|string> $condition */
+                foreach ($condition as $i => $operand) {
+                    $subCondition = $this->filterCondition($operand);
+                    if ($this->isEmpty($subCondition)) {
+                        unset($condition[$i]);
+                    } else {
+                        $condition[$i] = $subCondition;
+                    }
+                }
+
+                if (empty($condition)) {
+                    return [];
+                }
+
+                break;
+            case 'BETWEEN':
+            case 'NOT BETWEEN':
+                if (array_key_exists(1, $condition) && array_key_exists(2, $condition)) {
+                    if ($this->isEmpty($condition[1]) || $this->isEmpty($condition[2])) {
+                        return [];
+                    }
+                } else {
+                    return [];
+                }
+
+                break;
+            default:
+                if (array_key_exists(1, $condition) && $this->isEmpty($condition[1])) {
+                    return [];
+                }
+        }
+
+        array_unshift($condition, $operator);
+
+        return $condition;
     }
 
-    private function filterCondition(array $condition): array
+    /**
+     * Returns a value indicating whether the give value is "empty".
+     *
+     * The value is considered "empty", if one of the following conditions is satisfied:
+     *
+     * - it is `null`,
+     * - an empty string (`''`),
+     * - a string containing only whitespace characters,
+     * - or an empty array.
+     *
+     * @param mixed $value
+     *
+     * @return bool if the value is empty
+     */
+    private function isEmpty(mixed $value): bool
     {
-        return (array) $this->createQueryHelper()->filterCondition($condition);
+        return $value === '' || $value === [] || $value === null || (is_string($value) && trim($value) === '');
+    }
+
+    /**
+     * Normalizes format of ORDER BY data.
+     *
+     * @param array|ExpressionInterface|string $columns the columns value to normalize.
+     *
+     * See {@see orderBy} and {@see addOrderBy}.
+     */
+    private function normalizeOrderBy(array|string|ExpressionInterface $columns): array
+    {
+        if ($columns instanceof ExpressionInterface) {
+            return [$columns];
+        }
+
+        if (is_array($columns)) {
+            return $columns;
+        }
+
+        $columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
+        $result = [];
+
+        foreach ($columns as $column) {
+            if (preg_match('/^(.*?)\s+(asc|desc)$/i', $column, $matches)) {
+                $result[$matches[1]] = strcasecmp($matches[2], 'desc') ? SORT_ASC : SORT_DESC;
+            } else {
+                $result[$column] = SORT_ASC;
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Normalizes the SELECT columns passed to {@see select()} or {@see addSelect()}.
+     */
+    private function normalizeSelect(array|ExpressionInterface|string $columns): array
+    {
+        if ($columns instanceof ExpressionInterface) {
+            $columns = [$columns];
+        } elseif (!is_array($columns)) {
+            $columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
+        }
+
+        $select = [];
+
+        /** @psalm-var array<array-key, ExpressionInterface|string> $columns */
+        foreach ($columns as $columnAlias => $columnDefinition) {
+            if (is_string($columnAlias)) {
+                // Already in the normalized format, good for them.
+                $select[$columnAlias] = $columnDefinition;
+                continue;
+            }
+
+            if (is_string($columnDefinition)) {
+                if (
+                    preg_match('/^(.*?)(?i:\s+as\s+|\s+)([\w\-_.]+)$/', $columnDefinition, $matches) &&
+                    !preg_match('/^\d+$/', $matches[2]) &&
+                    !str_contains($matches[2], '.')
+                ) {
+                    /** Using "columnName as alias" or "columnName alias" syntax */
+                    $select[$matches[2]] = $matches[1];
+                    continue;
+                }
+                if (!str_contains($columnDefinition, '(')) {
+                    /** Normal column name, just alias it to itself to ensure it's not selected twice */
+                    $select[$columnDefinition] = $columnDefinition;
+                    continue;
+                }
+            }
+
+            // Either a string calling a function, DB expression, or sub-query
+            /** @var string */
+            $select[] = $columnDefinition;
+        }
+
+        return $select;
     }
 }
